Merge Week 3

+++ frontend/src/app/api/user/images/deleteImage.ts
@@ -1,26 +1,38 @@
 // Local Imports
-import { setupGridFS } from "@/lib/mongodb";
+import { getDB, setupGridFS } from "@/lib/mongodb";
 import { HTTP_RESPONSES } from "@/lib/constants/httpResponses";
-import { getValueFromSearchParams } from "@/lib/utils/routeHelpers";
 
 // Package Imports
 import { NextResponse } from "next/server";
 import { ObjectId } from "mongodb";
+import { isValidCollectionName } from "@/lib/utils/typeValidation/general";
+import { currentUser } from "@clerk/nextjs/server";
 
 export async function DELETE(req: Request) {
   try {
-    const imageId = getValueFromSearchParams(req, "id");
+    const { imageId, collectionName } = await req.json()
 
-    if (!imageId) {
+    console.log(imageId);
+    console.log(collectionName);
+
+    if (!imageId || !isValidCollectionName(collectionName)) {
       return NextResponse.json(
         { message: HTTP_RESPONSES.BAD_REQUEST },
         { status: 400 }
       );
     }
 
-    const { bucket } = await setupGridFS();
+    const { bucket, db } = await setupGridFS();
 
-    const imageIdAsObjectId = new ObjectId(imageId);
+    const clerkUser = await currentUser();
+
+    if (!clerkUser || !clerkUser.id) {
+      return NextResponse.json({ message: "User not found" }, { status: 404 });
+    }
+
+    const userProfile = await db.collection('users').findOne({ clerkId: clerkUser.id });
+
+    const imageIdAsObjectId = ObjectId.createFromHexString(imageId);
     const file = await bucket.find({ _id: imageIdAsObjectId }).toArray();
 
     if (!file.length || file.length === 0) {
@@ -30,7 +42,21 @@ export async function DELETE(req: Request) {
       );
     }
 
+    if (file[0].metadata?.creatorId.toString() !== userProfile?._id.toString()) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 403 });
+    }
+
+    const sourceId = file[0].metadata?.sourceId;
+
     await bucket.delete(file[0]._id);
+
+    await db.collection(collectionName).updateOne({
+      _id: ObjectId.createFromHexString(sourceId)
+    }, {
+      $set: {
+        imageId: null,
+      }
+    })
 
     return NextResponse.json({ message: HTTP_RESPONSES.OK }, { status: 200 });
   } catch (error) {
